requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
requests>=2.31.0
gspread>=6.0.0
google-auth>=2.0.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
openpyxl>=3.1.0
